Remove Array resize in Encoding::list
The output should be exactly EncodingCount items, reserve that space
beforehand and get rid of the array resize operations.

+++ include/natalie/encodings.hpp
@@ -2,7 +2,7 @@
 
 namespace Natalie {
 
-const int EncodingCount = 44;
+const size_t EncodingCount = 44;
 
 enum class Encoding {
     ASCII_8BIT = 1,
